Move management room to its own folder so we can start introspecting on it.
